+++ src/library/reeds_shepp/rs_path_interpolate.cc
@@ -27,8 +27,8 @@ int RSPathInterpolator::CopyRSPathKappaParam(RSPathKappaParam *dst_list,
 
 int RSPathInterpolator::CalcShortestRSPathKappa(
     RSPathKappaParam *kappa_list, const Pose2D *start_pose,
-    const Pose2D *goal_pose, double min_turn_radius,
-    const double inverse_radius, const RSPathRequestType request_type) {
+    const Pose2D *goal_pose, float min_turn_radius,
+    const float inverse_radius, const RSPathRequestType request_type) {
   bool is_same;
   RSPathParam *path;
   RSPathInfo *path_info = GetRSPathGlobalInfo();
@@ -77,8 +77,8 @@ int RSPathInterpolator::CalcShortestRSPathKappa(
 int RSPathInterpolator::CalcSCSPathKappa(RSPathKappaParam *kappa_list,
                                          const Pose2D *start_pose,
                                          const Pose2D *goal_pose,
-                                         double min_turn_radius,
-                                         const double inverse_radius,
+                                         float min_turn_radius,
+                                         const float inverse_radius,
                                          const RSPathRequestType request_type) {
   bool is_same;
   RSPathParam *path;
@@ -127,7 +127,7 @@ int RSPathInterpolator::CalcSCSPathKappa(RSPathKappaParam *kappa_list,
 
 int RSPathInterpolator::CalcRSPathKappa(RSPathKappaParam *kappa_list,
                                         const RSPathParam *path,
-                                        double inverse_radius) {
+                                        float inverse_radius) {
   int i;
 
   if (kappa_list == nullptr || path == nullptr) {
@@ -167,7 +167,7 @@ int RSPathInterpolator::CalcRSPathKappa(RSPathKappaParam *kappa_list,
 }
 
 int RSPathInterpolator::ShrinkRSPathByInvalidDist(RSPathKappaParam *kappa_list,
-                                                  double dist_thres) {
+                                                  float dist_thres) {
   int i, j;
 
   for (i = 0, j = 0; i < kappa_list->size; i++) {
@@ -189,7 +189,7 @@ int RSPathInterpolator::ShrinkRSPathByInvalidDist(RSPathKappaParam *kappa_list,
 int RSPathInterpolator::GetStraightLinePoint(RSPoint *goal_state,
                                              const RSPoint *start_state,
                                              AstarPathGear direction,
-                                             double length) {
+                                             float length) {
   if (direction == AstarPathGear::DRIVE) {
     goal_state->x = start_state->x + length * ifly_cos(start_state->theta);
     goal_state->y = start_state->y + length * ifly_sin(start_state->theta);
@@ -208,7 +208,7 @@ int RSPathInterpolator::GetStraightLinePoint(RSPoint *goal_state,
 int RSPathInterpolator::GetStraightLinePoint(RSPoint *goal_state,
                                              const RSPoint *start_state,
                                              AstarPathGear direction,
-                                             const double dist_to_start,
+                                             const float dist_to_start,
                                              const Pose2D *unit_vector) {
   goal_state->x = start_state->x + dist_to_start * unit_vector->x;
   goal_state->y = start_state->y + dist_to_start * unit_vector->y;
@@ -222,8 +222,8 @@ int RSPathInterpolator::GetStraightLinePoint(RSPoint *goal_state,
 
 int RSPathInterpolator::GetCirclePoint(RSPoint *goal_state,
                                        const RSPoint *start_state,
-                                       AstarPathGear direction, double length) {
-  double radius, theta;
+                                       AstarPathGear direction, float length) {
+  float radius, theta;
 
   radius = 1.0 / start_state->kappa;
   if (direction == AstarPathGear::DRIVE) {
@@ -248,7 +248,7 @@ int RSPathInterpolator::GetCirclePoint(RSPoint *goal_state,
 int RSPathInterpolator::InterpolateByKappa(RSPoint *next_point,
                                            const RSPoint *point,
                                            const AstarPathGear dir,
-                                           const double abs_len) {
+                                           const float abs_len) {
   if (!ifly_fequal(point->kappa, 0.0)) {
     GetCirclePoint(next_point, point, dir, abs_len);
   } else {
@@ -293,7 +293,7 @@ int RSPathInterpolator::UpdateAnchorPoint(RSAnchorPoints *point_set,
                                           const RSPathKappaParam *path_list) {
   RSPoint *cur_point, *next_point;
   int i;
-  double abs_length, length, kappa;
+  float abs_length, length, kappa;
   AstarPathGear dir;
 
   // update first anchor point
@@ -351,7 +351,7 @@ int RSPathInterpolator::UpdateAnchorPoint(RSAnchorPoints *point_set,
                                           const RSPath *path) {
   RSPoint *cur_point, *next_point;
   int i;
-  double abs_length, length, kappa;
+  float abs_length, length, kappa;
   AstarPathGear dir;
 
   // update first anchor point
@@ -405,10 +405,10 @@ int RSPathInterpolator::UpdateAnchorPoint(RSAnchorPoints *point_set,
 
 int RSPathInterpolator::RSPathSegmentInterpolate(
     RSPathSegment *path, const RSPoint *start_pose,
-    const RSSegmentKappaParam *path_kappa, double inc_dist) {
+    const RSSegmentKappaParam *path_kappa, float inc_dist) {
   RSPoint *state_curr, *state_next;
   int j, state_num;
-  double abs_length, length, s_seg, integration_step;
+  float abs_length, length, s_seg, integration_step;
   bool is_last;
 
   // update first point
@@ -468,10 +468,10 @@ int RSPathInterpolator::RSPathSegmentInterpolate(
 
 int RSPathInterpolator::PathSegmentInterpolateByLine(
     RSPathSegment *path, const RSPoint *start_pose, const RSPoint *next_pose,
-    const double path_kappa, const double length, const double inc_dist) {
+    const float path_kappa, const float length, const float inc_dist) {
   RSPoint *state_next;
   int j, state_num;
-  double abs_length, acc_s;
+  float abs_length, acc_s;
   bool is_last;
 
   // update first point
@@ -528,12 +528,12 @@ int RSPathInterpolator::PathSegmentInterpolateByLine(
 }
 
 int RSPathInterpolator::PathSegmentInterpolateByArc(
-    RSPathSegment *path, const RSPoint *start_pose, const double path_kappa,
-    const double length, const double inc_dist, const double radius,
-    const double inv_radius) {
+    RSPathSegment *path, const RSPoint *start_pose, const float path_kappa,
+    const float length, const float inc_dist, const float radius,
+    const float inv_radius) {
   RSPoint *state_next;
   int j, state_num;
-  double abs_length, acc_s;
+  float abs_length, acc_s;
   bool is_last;
 
   AstarPathGear gear = GetGearBySignLen(length);
@@ -554,7 +554,7 @@ int RSPathInterpolator::PathSegmentInterpolateByArc(
 
   // get vehicle circle
   VehicleCircle veh_circle;
-  double min_radius;
+  float min_radius;
   if (path_kappa > 0.0) {
     min_radius = radius;
   } else {
@@ -563,7 +563,7 @@ int RSPathInterpolator::PathSegmentInterpolateByArc(
 
   GetVehCircleByPose(&veh_circle, start_pose, min_radius, gear);
 
-  double length_eps = 0.001;
+  float length_eps = 0.001;
   for (j = 0; j < state_num; j++) {
     if (is_last) {
       break;
@@ -595,7 +595,7 @@ int RSPathInterpolator::PathSegmentInterpolateByArc(
 }
 
 int RSPathInterpolator::GetVehCircleByPose(VehicleCircle *veh_circle,
-                                           const RSPoint *pose, double radius,
+                                           const RSPoint *pose, float radius,
                                            AstarPathGear gear) {
   veh_circle->radius = radius;
   veh_circle->gear = gear;
@@ -609,9 +609,9 @@ int RSPathInterpolator::GetVehCircleByPose(VehicleCircle *veh_circle,
 int RSPathInterpolator::InterpolateByArcOffset(RSPoint *pose,
                                                const VehicleCircle *veh_circle,
                                                const RSPoint *start_pose,
-                                               const double arc,
-                                               const double inverse_radius) {
-  double delta_theta, theta;
+                                               const float arc,
+                                               const float inverse_radius) {
+  float delta_theta, theta;
 
   delta_theta = arc * inverse_radius;
 
@@ -630,7 +630,7 @@ int RSPathInterpolator::InterpolateByArcOffset(RSPoint *pose,
   // update next point theta
   theta = start_pose->theta + delta_theta;
 
-  double radius = veh_circle->radius;
+  float radius = veh_circle->radius;
 
   pose->x = veh_circle->center.x + radius * std::sin(theta);
   pose->y = veh_circle->center.y - radius * std::cos(theta);
